pyproject.toml
[project]
name = "mcp-langgraph-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain-mcp-adapters>=0.1.1",
    "langchain[openai]>=0.3.25",
    "langgraph>=0.4.5",
    "python-dotenv>=1.1.0",
]
